tools/mapmerge2/requirements.txt
pygit2==1.0.1
bidict==0.13.1
Pillow==6.2.0
